Add: LibraryCardコンポーネントを新規作成し、検索結果ページでライブラリ情報を表示
